+++ src/main/kotlin/com/example/backend/infrastructure/database/record/CustomerRecord.kt
@@ -1,7 +1,7 @@
 package com.example.backend.infrastructure.database.record
 
 data class CustomerRecord(
-    var customerId: Int? = null,
+    var customerId: Long? = null,
     var firstName: String? = null,
     var lastName: String? = null,
     var tenantName: String? = null
